feat(commands): update migration script
