--- alembic/versions/a8e9ee2cd4be_add_user_and_updated_date_info_to_.py
@@ -1,77 +1,0 @@
-"""add_user_and_updated_date_info_to_command_model
-
-Revision ID: a8e9ee2cd4be
-Revises: 00a9ceb38842
-Create Date: 2024-11-27 14:11:59.696512
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'a8e9ee2cd4be'
-down_revision = '00a9ceb38842'
-branch_labels = None
-depends_on = None
-
-# constraint name
-USER_ID_FKEY = 'commandblock_user_id_fk'
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table('commandblock', schema=None) as batch_op:
-        batch_op.add_column(sa.Column('user_id', sa.Integer(), nullable=True))
-        batch_op.add_column(sa.Column('updated_at', sa.DateTime(), nullable=True))
-        batch_op.alter_column('study_version',
-               existing_type=sa.VARCHAR(length=36),
-               nullable=True,
-               existing_server_default=sa.text("'fake_version'"))
-        batch_op.create_foreign_key(USER_ID_FKEY, 'identities', ['user_id'], ['id'], ondelete='CASCADE')
-
-    with op.batch_alter_table('group_metadata', schema=None) as batch_op:
-        batch_op.drop_index('ix_groupmetadatacopy_group_id')
-        batch_op.drop_index('ix_groupmetadatacopy_study_id')
-        batch_op.create_index(batch_op.f('ix_group_metadata_group_id'), ['group_id'], unique=False)
-        batch_op.create_index(batch_op.f('ix_group_metadata_study_id'), ['study_id'], unique=False)
-
-    with op.batch_alter_table('rawstudy', schema=None) as batch_op:
-        batch_op.drop_index('ix_rawstudycopy_missing')
-        batch_op.drop_index('ix_rawstudycopy_workspace')
-        batch_op.create_index(batch_op.f('ix_rawstudy_missing'), ['missing'], unique=False)
-        batch_op.create_index(batch_op.f('ix_rawstudy_workspace'), ['workspace'], unique=False)
-
-    with op.batch_alter_table('study_additional_data', schema=None) as batch_op:
-        batch_op.create_index(batch_op.f('ix_study_additional_data_patch'), ['patch'], unique=False)
-
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table('study_additional_data', schema=None) as batch_op:
-        batch_op.drop_index(batch_op.f('ix_study_additional_data_patch'))
-
-    with op.batch_alter_table('rawstudy', schema=None) as batch_op:
-        batch_op.drop_index(batch_op.f('ix_rawstudy_workspace'))
-        batch_op.drop_index(batch_op.f('ix_rawstudy_missing'))
-        batch_op.create_index('ix_rawstudycopy_workspace', ['workspace'], unique=False)
-        batch_op.create_index('ix_rawstudycopy_missing', ['missing'], unique=False)
-
-    with op.batch_alter_table('group_metadata', schema=None) as batch_op:
-        batch_op.drop_index(batch_op.f('ix_group_metadata_study_id'))
-        batch_op.drop_index(batch_op.f('ix_group_metadata_group_id'))
-        batch_op.create_index('ix_groupmetadatacopy_study_id', ['study_id'], unique=False)
-        batch_op.create_index('ix_groupmetadatacopy_group_id', ['group_id'], unique=False)
-
-    with op.batch_alter_table('commandblock', schema=None) as batch_op:
-        batch_op.drop_constraint(USER_ID_FKEY, type_='foreignkey')
-        batch_op.alter_column('study_version',
-               existing_type=sa.VARCHAR(length=36),
-               nullable=False,
-               existing_server_default=sa.text("'fake_version'"))
-        batch_op.drop_column('updated_at')
-        batch_op.drop_column('user_id')
-
-    # ### end Alembic commands ###
