+++ alembic/versions/bae9c99bc42d_add_user_name_and_updated_at_to_commands.py
@@ -1,0 +1,36 @@
+"""add_user_name_and_updated_at_to_commands
+
+Revision ID: bae9c99bc42d
+Revises: 00a9ceb38842
+Create Date: 2024-11-29 09:13:04.292874
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = 'bae9c99bc42d'
+down_revision = '00a9ceb38842'
+branch_labels = None
+depends_on = None
+
+USER_ID_FKEY = 'commandblock_user_id_fk'
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    with op.batch_alter_table('commandblock', schema=None) as batch_op:
+        batch_op.add_column(sa.Column('user_id', sa.Integer(), nullable=True))
+        batch_op.add_column(sa.Column('updated_at', sa.DateTime(), nullable=True))
+        batch_op.create_foreign_key(USER_ID_FKEY, 'identities', ['user_id'], ['id'], ondelete='CASCADE')
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    with op.batch_alter_table('commandblock', schema=None) as batch_op:
+        batch_op.drop_constraint(USER_ID_FKEY, type_='foreignkey')
+        batch_op.drop_column('updated_at')
+        batch_op.drop_column('user_id')
+    # ### end Alembic commands ###
